requirements.txt
streamlit>=1.35.0
opencv-python-headless>=4.9.0  # Streamlit Cloudでは headless版が必要
Pillow>=10.0.0
ultralytics>=8.0.0
numpy>=1.23.0
pandas>=2.0.0